refactor files, fix linting
