+++ backend/.eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  root: true,
+  env: { node: true, es2021: true },
+  extends: ['eslint:recommended', 'plugin:@stylistic/recommended-extends'],
+  rules: {
+    '@stylistic/semi': ['warn', 'always'],
+  },
+}
