+++ frontend/src/components/Home.jsx
@@ -1,4 +1,4 @@
-import React from 'react'
+
 
 
 const HomePage = () => {
